Structure: Python IDE
Structure is a Python IDE designed by a young teenager in 7 hours. It is actively being worked on, and will most likely crash if you type the wrong command. The IDE has a special feature: Scaffold, which is an automated code building tool that can help you to write tedious parts of your scripts.